Generalize PageTranslation model
- Add AbstractBasePageTranslation for attributes and methods shared with other page translation types
- Keep tree- and revision-related attributes and methods in the page translation model

(Prerequisite for #357)
