+++ src/cms/models/pages/__init__.py
@@ -1,7 +1,8 @@
 """
 This package contains page-related data models:
 
-* :class:`~cms.models.pages.abstract_base_page.AbstractBasePage`
+* :class:`~cms.models.pages.abstract_base_page.AbstractBasePage` and
+  :class:`~cms.models.pages.abstract_base_page_translation.AbstractBasePageTranslation`
 * :class:`~cms.models.pages.page.Page` and
   :class:`~cms.models.pages.page_translation.PageTranslation`
 """
